+++ resources/ts/Pages/SchoolOrder/Show.tsx
@@ -6,7 +6,6 @@ import { ProjectedOrders } from "@/Components/Sections/ProjectedOrdersSection";
 import SchoolSelectDropdown, { SchoolsList } from "@/Components/SchoolOrders/SchoolSelectDropdown";
 import ButtonLink from "@/Elements/Buttons/ButtonLink";
 import ChevronLeft from "@/Elements/Icons/ChevronLeft";
-import EditIcon from "@/Elements/Icons/EditIcon";
 
 import route from "ziggy-js";
 
@@ -19,12 +18,10 @@ export default function Show({ lessonOrder, schoolsList, classrooms, projectedOr
                 <ChevronLeft />{"Back to List"}
             </span></ButtonLink>
             <ContentWrapper title="School Lesson Order">
-                <div className="flex items-baseline gap-2 bg-gray-100 rounded ">
-
+                <div className="flex items-baseline gap-2 bg-gray-100 rounded">
                     <SchoolSelectDropdown currentSchoolId={lessonOrder.id} schoolsList={schoolsList} />
-                    <ButtonLink Icon={EditIcon} href={route('orders.edit', lessonOrder.id)}>Edit</ButtonLink>
                 </div>
-                <SchoolOrderSection lessonOrder={lessonOrder} classrooms={classrooms} projectedOrders={projectedOrders} curricula={curricula} />
+                <SchoolOrderSection classrooms={classrooms} projectedOrders={projectedOrders} curricula={curricula} schoolEmail={lessonOrder.email} />
 
             </ContentWrapper>
         </WrapperLayout>